+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added function `getFilterState` to get current status of filters
+- Added function `updateFilters` to update filters and refresh dataTable
+
 ## [3.4.0] - 2023-10-03
 
 ### Added